refactor/task: 주석 추가 및 타 도메인 서비스 연결
