+++ src/main/java/min/taskflow/task/controllder/TaskController.java
@@ -3,6 +3,7 @@ package min.taskflow.task.controllder;
 import lombok.RequiredArgsConstructor;
 import min.taskflow.common.response.ApiPageResponse;
 import min.taskflow.common.response.ApiResponse;
+import min.taskflow.task.dto.condition.TaskSearchCondition;
 import min.taskflow.task.dto.request.StatusUpdateRequest;
 import min.taskflow.task.dto.request.TaskCreateRequest;
 import min.taskflow.task.dto.request.TaskUpdateRequest;
@@ -10,11 +11,10 @@ import min.taskflow.task.dto.response.TaskResponse;
 import min.taskflow.task.service.commandService.ExternalCommandTaskService;
 import min.taskflow.task.service.queryService.ExternalQueryTaskService;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.awt.print.Pageable;
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -24,6 +24,7 @@ public class TaskController {
     private final ExternalCommandTaskService externalCommandTaskService;
     private final ExternalQueryTaskService externalQueryTaskService;
 
+    // 태스크 생성
     @PostMapping
     public ResponseEntity<ApiResponse<TaskResponse>> createTask(@RequestBody TaskCreateRequest request) {
 
@@ -43,6 +44,7 @@ public class TaskController {
         return ApiPageResponse.success(tasks, "Task 목록을 조회했습니다.");
     }
 
+    // 태스크 상세 조회
     @GetMapping("/{taskId}")
     public ResponseEntity<ApiResponse<TaskResponse>> getTaskByTaskId(@PathVariable Long taskId) {
 
@@ -51,25 +53,27 @@ public class TaskController {
         return ApiResponse.success(task, "Task를 조회했습니다.");
     }
 
+    // 태스크 상세 수정
     @PutMapping("/{taskId}")
     public ResponseEntity<ApiResponse<TaskResponse>> updateTaskDetailByTaskId(@PathVariable Long taskId,
-                                                                              TaskUpdateRequest taskUpdateRequest) {
+                                                                              @RequestBody TaskUpdateRequest taskUpdateRequest) {
 
         TaskResponse taskResponse = externalCommandTaskService.updateTaskDetailByTaskId(taskId, taskUpdateRequest);
 
         return ApiResponse.success(taskResponse, "Task가 수정되었습니다.");
     }
 
-    //Status 순차적 변경 적용
+    // 태스크 상태 수정
     @PatchMapping("/{taskId}/status")
     public ResponseEntity<ApiResponse<TaskResponse>> updateStatusByTaskId(@PathVariable Long taskId,
-                                                                          StatusUpdateRequest statusUpdateRequest) {
+                                                                          @RequestBody StatusUpdateRequest statusUpdateRequest) {
 
-        TaskResponse taskResponse = ExternalCommandTaskService.updateStatusByTaskId(taskId, statusUpdateRequest);
+        TaskResponse taskResponse = externalCommandTaskService.updateStatusByTaskId(taskId, statusUpdateRequest);
 
         return ApiResponse.success(taskResponse, "작업 상태가 업데이트되었습니다.");
     }
 
+    // 태스크 삭제(soft delete)
     @DeleteMapping("/{taskId}")
     public ResponseEntity<ApiResponse<Object>> deleteTaskByTaskId(@PathVariable Long taskId) {
