+++ src/main/java/min/taskflow/task/entity/Task.java
@@ -18,7 +18,7 @@ public class Task extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Long taskId;
 
     private String title;
