+++ src/main/java/min/taskflow/task/service/queryService/InternalQueryTaskService.java
@@ -1,16 +1,23 @@
 package min.taskflow.task.service.queryService;
 
 import lombok.RequiredArgsConstructor;
-import min.taskflow.task.mapper.TaskMapper;
-import min.taskflow.task.taskRepository.TaskRepository;
-import min.taskflow.user.service.InternalUserService;
+import min.taskflow.task.entity.Task;
+import min.taskflow.task.exception.TaskErrorCode;
+import min.taskflow.task.exception.TaskException;
+import min.taskflow.task.repository.TaskRepository;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class InternalQueryTaskService {
 
-    private final TaskMapper taskMapper;
-    private final InternalUserService internalUserService;
     private final TaskRepository taskRepository;
+
+    public Task findByTaskId(Long taskId) {
+
+        Task task = taskRepository.findById(taskId)
+                .orElseThrow(() -> new TaskException(TaskErrorCode.TASK_NOT_FOUND));
+
+        return task;
+    }
 }
